Ignore errors if the namespace deletion fails (#89)
We create a new namespace for each test scenario and delete the namespace once the test scenario execution gets completed. Sometimes the test scenario gets executed successfully but the test scenario will be marked as "Fail" if there is any failure in deletion of namespace. Hence, I created a new function which deletes the namespace and it does not expect command to get executed successfully